pyproject.toml
[project]
name = "mcp-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    # "faker>=37.5.3",
    "fastmcp>=2.11.2",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "langchain-groq>=0.3.7",
    "markdownify>=1.1.0",
    # "mcp[cli]>=1.12.3",
    "pdfplumber>=0.11.7",
    "pyngrok>=7.3.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "readabilipy>=0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "watchdog>=6.0.0",
]
